Added announcements API call
