+++ alabapp.xcodeproj/xcuserdata/janetcl.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,9 +14,9 @@
             filePath = "alabapp/API/ScheduleAPI.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "52"
-            endingLineNumber = "52"
-            landmarkName = "getData()"
+            startingLineNumber = "66"
+            endingLineNumber = "66"
+            landmarkName = "getSchedule()"
             landmarkType = "7">
             <Locations>
                <Location
