+++ alabapp/Components/AnnouncementsView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct AnnouncementsView: View {
+    @EnvironmentObject var network: Network
     let screenW = CGFloat((UIScreen.main.bounds.width - 48.0))
 
     var body: some View {
@@ -18,7 +19,7 @@ struct AnnouncementsView: View {
                 .foregroundColor(MyFont.black)
                 .frame(width: screenW, alignment: .topLeading)
             
-            Text("Sunday service on Sunday, June 14th will take place in Edison, NJ.\nContact us to confirm your attendance.")
+            Text(network.announcements.first?.values.message ?? "")
                 .padding(.horizontal, 6)
                 .font(MyFont.caption)
                 .foregroundColor(MyFont.black)
@@ -30,12 +31,14 @@ struct AnnouncementsView: View {
                         .frame(width: screenW, height: 70)
                         .shadow(color: .black.opacity(0.3), radius: 3, x: 2, y: 2)
                 )
+        }.onAppear {
+            network.getAnnouncements()
         }
     }
 }
 
 struct AnnouncementsView_Previews: PreviewProvider {
     static var previews: some View {
-        AnnouncementsView()
+        AnnouncementsView().environmentObject(Network())
     }
 }
